% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_common_elements.R
\name{count_common_elements}
\alias{count_common_elements}
\title{Count common SNPs}
\usage{
count_common_elements(vec1, vec2)
}
\arguments{
\item{vec1}{SNPs offspring}

\item{vec2}{SNPs parent 1 or parent 2}
}
\value{
common elements
}
\description{
Count the common SNPs between two vectors (offspring and parent 1 or parent 2)
}
\examples{
no example
}
