% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_common_markers.R
\name{estimate_common_markers}
\alias{estimate_common_markers}
\title{Common markers estimation}
\usage{
estimate_common_markers(marker_data, specific_lines_df)
}
\arguments{
\item{marker_data}{Hapmap with SNP markers information (letters)}

\item{specific_lines_df}{Dataframe with three columns 1) offspring, 2) parent 1, and 3) parent 2}
}
\value{
A dataframe with the number of common SNPs between offspring and parent
}
\description{
Estimate the number of common markers between offspring and parent 1 or parent 2
}
\examples{
estimate_common_markers(markers, crosses)
}
