% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parental_contribution.R
\name{parental_contribution}
\alias{parental_contribution}
\title{Parental contribution}
\usage{
parental_contribution(result_df)
}
\arguments{
\item{result_df}{The function use as input the output of the function "estimate_common_markers"}
}
\value{
A dataframe with the parental contribution in percentage between the offspring and parent 1 and parent 2
}
\description{
Parental contribution estimation is performed based on the formulae by Bernardo at al., 2000 titled Parental contribution and coefficient of coancestry among maize inbreds: pedigree, RFLP, and SSR data
}
\examples{
parental_contribution(results_df)
}
